requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
opencv-python
pillow
scipy
tqdm
easyocr 
opencv-python 
numpy